More fixes for #554
Once more unto the breach!  Fixed standard font array to be used across the entire class.  This was necessary for matching against Google font strings and weeding out the standard font when setting up the Google font css call.  Also integrated the font-backup to append to family-font in output and compiler.